+++ lib/state/market-creation/types/form.ts
@@ -1,7 +1,11 @@
-import * as O from "@zeitgeistpm/utility/dist/option";
-import { ChainTime, dateBlock } from "@zeitgeistpm/utility/dist/time";
+import {
+  CreateMarketParams,
+  MetadataStorage,
+  RpcContext,
+} from "@zeitgeistpm/sdk-next";
 import { BLOCK_TIME_SECONDS } from "lib/constants";
 import moment from "moment";
+import { DeepRequired } from "react-hook-form";
 import * as z from "zod";
 import {
   IOAnswers,
@@ -21,7 +25,6 @@ import {
   IOTags,
   IOYesNoAnswers,
 } from "./validation";
-import { DeepRequired } from "react-hook-form";
 
 /**
  * This is the type of the full market creation form data that is used to create a market.
@@ -92,54 +95,13 @@ export type Moderation = z.infer<typeof IOModerationMode>;
 export type Liquidity = z.infer<typeof IOLiquidity>;
 export type LiquidityRow = z.infer<typeof IOLiquidityRow>;
 
-export type BlockTimeline = {
-  market: { end: number };
-  grace: { period: number; end: number };
-  report: { period: number; end: number };
-  dispute: { period: number; end: number };
-};
-
-export const timelineAsBlocks = (
-  periods: {
-    marketEndDate?: Date;
-    gracePeriod: Partial<PeriodOption>;
-    reportingPeriod: Partial<PeriodOption>;
-    disputePeriod: Partial<PeriodOption>;
-  },
-  chainTime?: ChainTime,
-): O.IOption<BlockTimeline> => {
-  return O.tryCatch(() => {
-    if (!chainTime) return null;
-
-    const marketEndDate = new Date(periods.marketEndDate);
-    const marketEndBlock = dateBlock(chainTime, marketEndDate);
-
-    const gracePeriodEndBlock =
-      periods.gracePeriod?.type === "date"
-        ? periods.gracePeriod?.block
-        : marketEndBlock + durationasBlocks(periods.gracePeriod);
-
-    const reportPeriodEndBlock =
-      periods.reportingPeriod?.type === "date"
-        ? periods.reportingPeriod?.block
-        : gracePeriodEndBlock + durationasBlocks(periods.reportingPeriod);
-
-    const disputePeriodEndBlock =
-      periods.disputePeriod?.type === "date"
-        ? periods.disputePeriod?.block
-        : reportPeriodEndBlock + durationasBlocks(periods.disputePeriod);
-
-    const graceDelta = gracePeriodEndBlock - marketEndBlock;
-    const reportDelta = reportPeriodEndBlock - gracePeriodEndBlock;
-    const disputeDelta = disputePeriodEndBlock - reportPeriodEndBlock;
-
-    return {
-      market: { end: marketEndBlock },
-      grace: { period: graceDelta, end: gracePeriodEndBlock },
-      report: { period: reportDelta, end: reportPeriodEndBlock },
-      dispute: { period: disputeDelta, end: disputePeriodEndBlock },
-    };
-  }).bind(O.fromNullable);
+/**
+ * Create a the needed params for the market creation extrinsic from the form data.
+ */
+export const marketFormDataToExtrinsicParams = (
+  form: ValidMarketFormData,
+): CreateMarketParams<RpcContext<MetadataStorage>, MetadataStorage> => {
+  throw new Error("Not implemented");
 };
 
 export const durationasBlocks = (duration: Partial<PeriodDurationOption>) => {
